% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openseapalette.R
\name{add_logo}
\alias{add_logo}
\title{Arrange logo on graph}
\usage{
add_logo(p1, ...)
}
\arguments{
\item{p1}{ggplot2 plot to add logo on}

\item{...}{Additional arguments}
}
\description{
Arrange logo on graph
}
