% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mayc_avg_price}
\alias{mayc_avg_price}
\title{Mutant Ape Yacht Club Average Price}
\format{
A data frame with 227 entries, one entry for each day:
\describe{
 \item{date}{Date}
 \item{avgprice}{Average price of asset within MAYC collection}
 \item{volume}{Quantity of ETH traded during that day}
 \item{numsales}{Number of assets sold during that day}
}
}
\usage{
mayc_avg_price
}
\description{
Mutant Ape Yacht Club Average Price
}
\keyword{datasets}
