% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openseapalette.R
\name{opensea_cols}
\alias{opensea_cols}
\title{Function to extract OpenSea colors as hex codes}
\usage{
opensea_cols(...)
}
\arguments{
\item{...}{Character names of opensea_colors}
}
\description{
Function to extract OpenSea colors as hex codes
}
