% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openseapalette.R
\name{opensea_logo}
\alias{opensea_logo}
\title{Place logo PNG}
\usage{
opensea_logo(filename, ...)
}
\arguments{
\item{filename}{File name of logo in PNG format}

\item{...}{Additional arguments}
}
\description{
Place logo PNG
}
