% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openseapalette.R
\name{opensea_pal}
\alias{opensea_pal}
\title{Return function to interpolate an OpenSea color palette}
\usage{
opensea_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in OpenSea color palette}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate an OpenSea color palette
}
