% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensea.R
\name{read_os_collection}
\alias{read_os_collection}
\title{Retrieve data from OpenSea}
\source{
OpenSea
}
\usage{
read_os_collection(
  queryString = list(offset = "0", limit = "300"),
  api_key = NULL,
  ...
)
}
\arguments{
\item{queryString}{a list of options}

\item{api_key}{OpenSEA API key}

\item{...}{currently ignored}
}
\description{
Retrieve data from OpenSea
}
\examples{
read_os_collection()
read_os_collection(list(limit = "20"))
read_os_collection(api_key = "6dd0b79cbb934d72853ad157a14f78ca")
}
\references{
https://docs.opensea.io/reference/collection-model
}
