% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openseapalette.R
\docType{data}
\name{opensea_colors}
\alias{opensea_colors}
\title{References:
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
OpenSea official colors}
\format{
An object of class \code{character} of length 4.
}
\usage{
opensea_colors
}
\description{
References:
https://drsimonj.svbtle.com/creating-corporate-colour-palettes-for-ggplot2
OpenSea official colors
}
\keyword{datasets}
